project structure changes
